Component updates
Turned farmlandmanager into a component so it can be used in the player object

Turned PlantedCrop  into a component as this can be used in the plant object which is a child of the crop plot.

Added logic to the player input controller to allow the player to place farmland and plant wheat seeds.